fix: fix previous commit
The development server was not working at all after the last commit.
It turns out that a `style.js` was being emitted which was causing all sorts of problems.
This is now fixed by preventing it from being emitted with the `ignore-emit-webpack-plugin` plugin as well as some other changes found in GitHub issues.

I've also removed the `chunk.` prefix, it's not something meaningful at all.